Use better names for snippets
