+++ apps/docs/src/guide/transactions/snippets/assemble-tx/get-cost-and-fund-new.ts
@@ -1,28 +1,24 @@
 import { Provider, ScriptTransactionRequest, Wallet } from 'fuels';
 
-import {
-  LOCAL_NETWORK_URL,
-  WALLET_PVT_KEY,
-  WALLET_PVT_KEY_2,
-} from '../../../../env';
+import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../../../../env';
 
 const provider = new Provider(LOCAL_NETWORK_URL);
 const accountA = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
-const accountB = Wallet.fromPrivateKey(WALLET_PVT_KEY_2, provider);
+const accountB = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
 const baseAssetId = await provider.getBaseAssetId();
 
-// #region assemble-tx-2
-const request = new ScriptTransactionRequest();
+// #region get-cost-and-fund-new
+const transferAmount = 100;
 
-// Add a coin output to transfer 100 base asset to accountB
-request.addCoinOutput(accountB.address, 100, baseAssetId);
+const request = new ScriptTransactionRequest();
+request.addCoinOutput(accountB.address, transferAmount, baseAssetId);
 
 const { assembledRequest } = await provider.assembleTx({
   request,
   feePayerAccount: accountA,
   accountCoinQuantities: [
     {
-      amount: 100,
+      amount: transferAmount,
       assetId: baseAssetId,
       account: accountA,
       changeOutputAccount: accountA,
@@ -31,5 +27,7 @@ const { assembledRequest } = await provider.assembleTx({
 });
 
 const tx = await accountA.sendTransaction(assembledRequest);
-await tx.waitForResult();
-// #endregion assemble-tx-2
+const { isStatusSuccess } = await tx.waitForResult();
+
+// #endregion get-cost-and-fund-new
+console.log('isStatusSuccess', isStatusSuccess);
